cs50w_final_project/backend/requirements.txt
Django
django-quill-editor
Pillow